fix: change js-interact to ts-interact
