--- js-interact/index.ts
@@ -1,4 +1,0 @@
-import JsInteract from './JsInteract'
-
-
-new JsInteract(3000);
