+++ ts-interact/index.ts
@@ -1,0 +1,4 @@
+import JsInteract from './TsInteract'
+
+
+new JsInteract(3000);
